+++ algo/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 * Fixed crate name.
 
+* Strongly connected components.
+
 ## [0.1.0] - 2025-03-31
 
 ### New